Initial commit
Gathered all pet projects and put them all together in a mono-repo.
As the number of pet projects grew, it became more and more tedious to maintain common code. At first I tried to isolate the common code into a separate repo, but it was still a pain to maintain, using git submodules or CMake's FetchContent.
Finally I decided to put everything in a single repo, just for practical reasons.
Since this repo contains "all kinds of animals", I chose to name the main project 'zoo'.